+++ test/resources/language/emacs-lisp/checkdoc-headerless.el
@@ -1,0 +1,3 @@
+(defun flycheck-headerless-snippet (x)
+  "Frob X"
+  x)